requirements.txt
BeautifulSoup==3.2.1
click==3.3
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.2
django-cors-headers==1.1.0
django-filter==0.9.2
django-toolbelt==0.0.1
djangorestframework==3.1.0
gunicorn==19.3.0
psycopg2==2.6
redis==2.10.3
requests==2.5.3
rq==0.5.0
static3==0.5.1
